add prisma db setup, saves user id and email for future kindel sendings

+++ src/helpers/createBookInfoMenu.ts
@@ -1,8 +1,7 @@
 import { MenuRange } from '@grammyjs/menu'
-import { InputFile } from 'grammy'
 import { BotContext } from '..'
-import { IBook } from '../types/book'
 import { downloadBook } from '../utils/downloadBook'
+import { getExtension } from '../utils/getExtension'
 import { backToMain } from './backToMainMenu'
 
 export const createBookInfoMenu = (
@@ -11,21 +10,17 @@ export const createBookInfoMenu = (
 ) => {
   range.submenu(ctx.t('download'), 'download-menu').row()
   if (
+    ctx.session.book &&
     ctx.session.book.downloadLinks.filter(
       (link) => link === 'download' || 'epub'
     ).length > 1
   ) {
-    ctx.session.bookFileExtension = ctx.session.book.downloadLinks.includes(
-      'epub'
-    )
-      ? 'epub'
-      : 'pdf'
     range
       .text(ctx.t('book_sent_to_kindle'), (ctx) => {
-        downloadBook(
-          `${ctx.session.book.href}/${ctx.session.bookFileExtension}`,
-          ctx
-        ).then((res) => {
+        const ext = ctx.session.book.downloadLinks.includes('epub')
+          ? 'epub'
+          : 'pdf'
+        downloadBook(`${ctx.session?.book?.href}/${ext}`, ctx).then((res) => {
           ctx.deleteMessage()
           ctx.scenes.enter('send-book')
         })
